refactor: Page rename
- PageName > PageTitle
